Added SpyObject. Added test object explanation to TestDefinitions.h
