+++ SpatialGDK/Source/SpatialGDKTests/Private/DummyObject.h
@@ -8,15 +8,10 @@
 
 #include "DummyObject.generated.h"
 
-TArray<uint8> TypeToArray(ESchemaComponentType Type);
-ESchemaComponentType ArrayToType(const TArray<uint8>& Array);
-
 UCLASS()
 class UDummyObject : public UObject
 {
 	GENERATED_BODY()
 public:
 	bool ProcessRPC(const FPendingRPCParams& Params);
-
-	TMap<ESchemaComponentType, TArray<uint32>> ProcessedRPCIndices;
 };
